v0.4.0
added:
photos
DTO

changes:
http request
services
less mysql requests

+++ src/main/java/com/Maruszak/MantisKeeper/services/InvertebratesServiceImpl.java
@@ -1,5 +1,6 @@
 package com.Maruszak.MantisKeeper.services;
 
+import com.Maruszak.MantisKeeper.DTO.InvertDTO;
 import com.Maruszak.MantisKeeper.model.Instar;
 import com.Maruszak.MantisKeeper.model.Invertebrate;
 import com.Maruszak.MantisKeeper.model.User;
@@ -9,6 +10,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
+import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.time.LocalDate;
@@ -28,6 +30,9 @@ public class InvertebratesServiceImpl {
     @Autowired
     private UserDetailsServiceImpl userService;
 
+    @Autowired
+    private PhotoServiceImpl photoService;
+
 
     public Invertebrate findInvertById(UUID id) {
         Optional<Invertebrate> invertOptional = invertRepository.findById(id);
@@ -38,69 +43,88 @@ public class InvertebratesServiceImpl {
     }
 
     @Transactional
-    public void saveInvertAndInstar(Instar instar) {
-        instarService.addInstar(instar);
-        Invertebrate invertebrate = instar.getInvertebrate();
-        invertRepository.save(invertebrate);
+    public void saveNewInvert(InvertDTO invertDTO, List<MultipartFile> photos) {
+        invertDTO.getInvertebrate().setUser(userService.getUser());
+        for (Instar instar : invertDTO.getInstars()) {
+            instar.setInvertebrate(invertDTO.getInvertebrate());
+        }
+        instarService.saveAllInstars(invertDTO.getInstars());
+        invertRepository.save(invertDTO.getInvertebrate());
+        photoService.saveFiles(photos, invertDTO.getInvertebrate());
     }
 
     @Transactional
-    public void updateInvertAndInstar(List<Instar> instars) {
-        instarService.deleteAllByInvert(instars.get(0).getInvertebrate());
-        for (Instar instar : instars) {
-            instarService.saveInstar(instar);
+    public void updateInvert(InvertDTO invertDTO, List<MultipartFile> photos, List<UUID> ids) {
+        User user = userService.getUser();
+        invertDTO.getInvertebrate().setUser(user);
+        for (Instar instar : invertDTO.getInstars()) {
+            instar.setInvertebrate(invertDTO.getInvertebrate());
         }
-        invertRepository.save(instars.get(0).getInvertebrate());
+        instarService.deleteAllByInvert(invertDTO.getInvertebrate());
+        instarService.saveAllInstars(invertDTO.getInstars());
+        invertRepository.save(invertDTO.getInvertebrate());
+        photoService.deleteAllById(ids);
+        photoService.saveFiles(photos, invertDTO.getInvertebrate());
     }
 
     @Transactional
     public void deleteInvert(UUID id) {
+        User user = userService.getUser();
         Invertebrate invert = findInvertById(id);
+        if(!invert.getUser().getId().equals(user.getId())){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Not Your invert cannot delete");
+        }
         instarService.deleteAllByInvert(invert);
+        photoService.deleteAllByInvert(invert);
         invertRepository.deleteById(id);
     }
 
     @Transactional
-    public void saveAsDead(UUID id , LocalDate date) {
-        Invertebrate invertebrate = findInvertById(id);
-        invertebrate.setAlive(false);
-        invertebrate.setDeath(date);
-        invertRepository.save(invertebrate);
+    public void saveAsDead(UUID id, LocalDate date) {
+        User user = userService.getUser();
+        Invertebrate invert = findInvertById(id);
+        if(!invert.getUser().getId().equals(user.getId())){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Not Your invert, cannot change it");
+        }
+        invert.setAlive(false);
+        invert.setDeath(date);
+        invertRepository.save(invert);
     }
 
     public String editInvertHTML(UUID id, Model model) {
         User user = userService.getUser();
-        List<Invertebrate> invertebrates = invertRepository.findAllByUser(user);
-        for(Invertebrate invert : invertebrates){
-            invert.setInstars(instarService.findInstarsByInvertAsc(invert));
-        }
-        user.setInvertebratesList(invertebrates);
         Invertebrate invert = findInvertById(id);
+        if(!invert.getUser().getId().equals(user.getId())){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Not Your invert, cannot edit");
+        }
+        invert.setUser(user);
+        invert.setInstars(instarService.findInstarsByInvertAsc(invert));
+        invert.setPhotos(photoService.findAllByInvert(invert));
         model.addAttribute("invert", invert);
-        model.addAttribute("user", user);
+
         return "editInvert";
     }
 
     public String markInvertDeadHTML(UUID id, Model model) {
-        model.addAttribute("id" , id);
+        model.addAttribute("id", id);
         return "markDead";
     }
 
     public String deadInvertsHTML(Model model) {
         User user = userService.getUser();
         List<Invertebrate> invertebrates = invertRepository.findAllByUserAndAliveFalse(user);
-        for(Invertebrate invert : invertebrates){
+        for (Invertebrate invert : invertebrates) {
             invert.setInstars(instarService.findInstarsByInvertAsc(invert));
         }
         user.setInvertebratesList(invertebrates);
-        model.addAttribute("user" , user);
+        model.addAttribute("user", user);
         return "deadInverts";
     }
 
     public String getMyAliveInvertsHTML(Model model) {
         User user = userService.getUser();
         List<Invertebrate> invertebrates = invertRepository.findAllByUserAndAliveTrue(user);
-        for(Invertebrate invert : invertebrates){
+        for (Invertebrate invert : invertebrates) {
             invert.setInstars(instarService.findInstarsByInvertAsc(invert));
         }
         user.setInvertebratesList(invertebrates);
@@ -110,13 +134,18 @@ public class InvertebratesServiceImpl {
 
     public String addInvertHTML(Model model) {
         User user = userService.getUser();
+        List<Invertebrate> invertebrates = invertRepository.findAllByUser(user);
+        user.setInvertebratesList(invertebrates);
         model.addAttribute("user", user);
         return "addInvert";
     }
 
     public String invertDetailsHTML(UUID id, Model model) {
         Invertebrate invert = findInvertById(id);
+        invert.setPhotos(photoService.findAllByInvert(invert));
+        invert.setInstars(instarService.findInstarsByInvertAsc(invert));
         model.addAttribute("invert", invert);
         return "invertDetails";
     }
+
 }
